Link menu up/down navigation to fx dial
Originally I wanted to have this do up/down difficulty levels, but there is no dedicated key for that. This is a compromise and also lets the user navigate song selection. As an added bonus, this means that with +/- and the effect dial the full menu can be navigated even without using the joystick!

This also adds a 'get' method to the HID button helper class, so that the effects dial does not interfere with the joystick's outputs normally.

The timeout and threshold levels may need to be tweaked later - the fx potentiometer on the testing turntable is a little broken and spits out bad values on occassion.