Revert "Increase I2C bus speed to 400 KHz"
To wit - if this *shouldn't* matter, then why do it? Increasing the bus speed also increases the possibility of errors in the data stream and doesn't gain anything since the controls program is so lightweight. Reverting and keeping it in the commit history because I may add it back later, but for now we're staying at the default speed.

This reverts commit 4beac4ef19dfe4edce487733ef5d20ce385ec797.