Print Serial version on 32U4 boards
Previously this was only checking for Teensy USB modes. As the 32U4 boards under Arduino always have CDC serial, this is a catch-all.

In the future it would be better to make my own define for AVR boards with the built-in CDC definition, but as only 32U4 and Teensy boards are supported this will do fine for now.